+++ lib/openapi_first/test.rb
@@ -8,9 +8,20 @@ module OpenapiFirst
   module Test
     autoload :Coverage, 'openapi_first/test/coverage'
     autoload :Methods, 'openapi_first/test/methods'
+    autoload :Callable, 'openapi_first/test/callable'
     extend Registry
 
     class CoverageError < Error; end
+
+    # @visible private
+    module Observed; end
+
+    # Inject request/response validation in a rack app class
+    def self.observe(app, api: :default)
+      mod = Callable[api:]
+      app.prepend(mod) unless app.include?(Observed)
+      app.include(Observed)
+    end
 
     def self.minitest?(base)
       base.include?(::Minitest::Assertions)
@@ -31,9 +42,9 @@ module OpenapiFirst
 
       install
       yield configuration
-      configuration.registry.each do |name, oad|
-        register(oad, as: name)
-      end
+
+      configuration.registry.each { |name, oad| register(oad, as: name) }
+      configuration.apps.each { |name, app| observe(app, api: name) }
       Coverage.start(skip_response: configuration.skip_response_coverage)
 
       if definitions.empty?
@@ -61,9 +72,10 @@ module OpenapiFirst
       )
       return unless configuration.report_coverage == true
 
-      return if Coverage.result.coverage >= configuration.minimum_coverage
+      coverage = Coverage.result.coverage
+      return if coverage >= configuration.minimum_coverage
 
-      puts 'API Coverage fails with exit 2, because not all described requests and responses have been tested.'
+      puts "API Coverage fails with exit 2, because not all described requests and responses have been tested (#{coverage.round(4)}% covered)." # rubocop:disable Layout/LineLength
 
       exit 2
     end