+++ lib/openapi_first/test/configuration.rb
@@ -13,15 +13,23 @@ module OpenapiFirst
         @ignored_unknown_status = [404]
         @report_coverage = true
         @registry = {}
+        @apps = {}
       end
 
       # Register OADs, but don't load them just yet
+      # @param [OpenapiFirst::OAD] oad The OAD to register
+      # @param [Symbol] as The name to register the OAD under
       def register(oad, as: :default)
         @registry[as] = oad
       end
 
+      # Observe a rack app
+      def observe(app, api: :default)
+        @apps[api] = app
+      end
+
       attr_accessor :coverage_formatter_options, :coverage_formatter, :response_raise_error
-      attr_reader :registry, :minimum_coverage, :report_coverage, :ignored_unknown_status
+      attr_reader :registry, :apps, :minimum_coverage, :report_coverage, :ignored_unknown_status
 
       def minimum_coverage=(value)
         warn 'Setting OpenapiFirst::Test::Configuration#minimum_coverage= is deprecated ' \
@@ -30,6 +38,8 @@ module OpenapiFirst
         @minimum_coverage = value
       end
 
+      # Configure report coverage
+      # @param [Boolean, :warn] value Whether to report coverage or just warn.
       def report_coverage=(value)
         allowed_values = [true, false, :warn]
         unless allowed_values.include?(value)